[SQL][PYTHON] Move aggregator module to pyspark.sql and set version to 4.4.0
Relocate `aggregator.py` from `pyspark.sql.pandas` to `pyspark.sql` (import as
`pyspark.sql.aggregator`), and set the `versionadded` for `Aggregator`/`udaf`
to 4.4.0. Update the references in util.py, connect/udf.py, and the test.
