+++ python/pyspark/sql/connect/udf.py
@@ -210,7 +210,7 @@ class UserDefinedFunction:
             eval_type=self.evalType,
             func=self.func,
             python_ver="%d.%d" % sys.version_info[:2],
-            # Set for incremental Python aggregators (see pyspark.sql.pandas.aggregator).
+            # Set for incremental Python aggregators (see pyspark.sql.aggregator).
             buffer_type=getattr(self, "bufferSchema", None),
         )
         return CommonInlineUserDefinedFunction(
